feature(scribe): add support for thrift compression
This commit adds thrift compression support on the server and
network store. Add thrift_compression=on to your configuration file
to enable server compression. Add remote_thrift_compression=on
to the network store section of the configuration file for which
you wish to enable compression.

Compressed and uncompressed network stores can be used at the same
time, but the server will only accept either compressed or
uncompressed client connections, as thrift doesn't support
negotiation at this time.

The goal is to enable compression on cross DC connections.

This needs thrift > 0.9 as it introduces TZlibTransportFactory.